+++ src/DictVectors/abstractdvec.jl
@@ -1,3 +1,6 @@
+###
+### This file contains methods defined on `AbstractDVec`
+###
 """
     DictVectors.AbstractDVec{K,V}
 
@@ -13,7 +16,7 @@ number of stored elements, not the size of the vector space.
 
 They have a [`StochasticStyle`](@ref) which selects the spawning algorithm in `FCIQMC`.
 
-To iterate over an `AbstractDVec`, use `pairs` or `values`.
+To iterate over an `AbstractDVec`, use `keys`, `pairs`, or `values`.
 
 # Interface
 
@@ -22,8 +25,9 @@ The interface is similar to the `AbstractDict` interface.
 Implement what would be needed for the `AbstractDict` interface (`pairs`, `keys`, `values`,
 `setindex!, getindex, delete!, length, haskey, empty!, isempty`) and, in addition:
 * [`StochasticStyle`](@ref)
-* [`storage(dv)`](@ref) returns an `AbstractDict` storing the raw data with possibly
+* [`storage`](@ref) returns an `AbstractDict` storing the raw data with possibly
   different `valtype` than `V`.
+* [`deposit`](@ref)
 """
 abstract type AbstractDVec{K,V} end
 
@@ -39,25 +43,6 @@ function Base.show(io::IO, dvec::AbstractDVec)
         end
     end
 end
-
-"""
-    deposit!(w::AbstractDVec, add, val, parent::Pair)
-
-Add `val` into `w` at address `add`, taking into account initiator rules if applicable.
-`parent` contains the `address => value` pair from which the pair `add => val`
-was created. [`InitiatorDVec`](@ref) can intercept this and add its own functionality.
-"""
-function deposit!(w, add, val, _)
-    w[add] += convert(valtype(w), val)
-end
-
-"""
-    storage(dvec) -> AbstractDict
-
-Return the raw storage associated with `dvec` as an `AbstractDict`. Used in MPI
-communication.
-"""
-storage
 
 ###
 ### Types
@@ -75,14 +60,7 @@ Base.ndims(::AbstractDVec) = 1
 ###
 ### copy*, zero*
 ###
-"""
-    zero!(v)
-
-Replace `v` by a zero vector as an inplace operation. For `AbstractDVec` types it means
-removing all non-zero elements. For `AbstractArrays`, it sets all of the values to zero.
-"""
 zero!(v::AbstractDVec) = empty!(v)
-zero!(v::AbstractVector{T}) where {T} = v .= zero(T)
 
 Base.zero(dv::AbstractDVec) = empty(dv)
 
@@ -332,13 +310,6 @@ walkernumber(w) = walkernumber(StochasticStyle(w), w)
 walkernumber(::StochasticStyle, w) = dot(Norm1ProjectorPPop(), w)
 # complex walkers as two populations
 # the following default is fast and generic enough to be good for real walkers and
-
-"""
-    localpart(dv) -> AbstractDVec
-
-Get the part of `dv` that is located on this MPI rank. Returns `dv` itself for `DictVector`s.
-"""
-localpart(dv) = dv # default for local data
 
 """
     FrozenDVec
