+++ src/Hamiltonians/BoseHubbardReal1D2C.jl
@@ -48,7 +48,7 @@ function starting_address(h::BoseHubbardReal1D2C)
     return BoseFS2C(starting_address(h.ha), starting_address(h.hb))
 end
 
-LOStructure(::Type{<:BoseHubbardReal1D2C{<:Real}}) = Hermitian()
+LOStructure(::Type{<:BoseHubbardReal1D2C{<:Real}}) = IsHermitian()
 
 Base.getproperty(h::BoseHubbardReal1D2C, s::Symbol) = getproperty(h, Val(s))
 Base.getproperty(h::BoseHubbardReal1D2C, ::Val{:ha}) = getfield(h, :ha)
